requirements.txt
python-wordpress-xmlrpc
python-slugify